refactor: extract placement analysis service
