+++ scripts/release_artifact_allowlist.txt
@@ -219,10 +219,15 @@ src/diptrace_mcp/service.py
 src/diptrace_mcp/services/__init__.py
 src/diptrace_mcp/services/bom.py
 src/diptrace_mcp/services/context.py
+src/diptrace_mcp/services/discovery.py
 src/diptrace_mcp/services/documents.py
+src/diptrace_mcp/services/exports.py
 src/diptrace_mcp/services/external_jobs.py
+src/diptrace_mcp/services/jobs.py
+src/diptrace_mcp/services/placement.py
 src/diptrace_mcp/services/review.py
 src/diptrace_mcp/services/routing.py
+src/diptrace_mcp/services/semantic_operations.py
 src/diptrace_mcp/sessions.py
 src/diptrace_mcp/silkscreen.py
 src/diptrace_mcp/spatial.py
